requirements.txt
requests == 2.32.3
python-dotenv==1.0.1
openpyxl == 3.1.5
schedule==1.2.2
slack_sdk==3.34.0
